Sample EmailPolicy APK sources
Create a sample (skeleton) package for an Email policy module.  Can be
used on Froyo or Gingerbread.

These policy files are intended for use by an OEM who is building a
device using the Android Open Source platform, but wishes to provide small
customizations to the Email application without having to edit or merge
changes directly into its source.  It also allows a single Email
application APK to be deployed to different devices with slightly
different policies.

Change-Id: I4e7e103dbb9bc257f2bd439eb508a51a8da853ba