Set the protocol for T-Mobile USA to IPv6.
This will cause new or factory-reset phones to connect to
T-Mobile using IPv6 and 464xlat, allowing us to droidfood and
dogfood the 464xlat code.

IPv6 on T-Mobile was successfully tested on JB MR2 on Galaxy
Nexus, Nexus 4, and Nexus 7 3G. We couldn't test deb since
we haven't been able to find anyone with one yet.

Bug: 8276725
Change-Id: I68ffd3af00c82f56665d429a7c719b8908db8189